# Optimized issue_query